+++ skforecast/ForecasterAutoregMultiVariate/tests/test_fit.py
@@ -96,7 +96,7 @@ def test_forecaster_index_step_stored():
     assert results == expected
     
 
-@pytest.mark.parametrize("n_jobs", [1, -1], 
+@pytest.mark.parametrize("n_jobs", [1, -1, 'auto'], 
                          ids=lambda n_jobs: f'n_jobs: {n_jobs}')
 def test_fit_in_sample_residuals_stored(n_jobs):
     """
@@ -105,9 +105,9 @@ def test_fit_in_sample_residuals_stored(n_jobs):
     series = pd.DataFrame({'l1': pd.Series(np.arange(10)), 
                            'l2': pd.Series(np.arange(10))})
     
-    forecaster = ForecasterAutoregMultiVariate(LinearRegression(), 
-                                               level='l1', lags=3, steps=2)
-    forecaster.fit(series=series, n_jobs=n_jobs)
+    forecaster = ForecasterAutoregMultiVariate(LinearRegression(), level='l1', 
+                                               lags=3, steps=2, n_jobs=n_jobs)
+    forecaster.fit(series=series)
     expected = {1: np.array([0.0000000e+00, 0.0000000e+00, 0.0000000e+00, 
                              0.0000000e+00, 0.0000000e+00, 8.8817842e-16]),
                 2: np.array([0., 0., 0., 0., 0., 0.])}
@@ -119,7 +119,7 @@ def test_fit_in_sample_residuals_stored(n_jobs):
     assert all(all(np.isclose(results[k], expected[k])) for k in expected.keys())
 
 
-@pytest.mark.parametrize("n_jobs", [1, -1], 
+@pytest.mark.parametrize("n_jobs", [1, -1, 'auto'], 
                          ids=lambda n_jobs: f'n_jobs: {n_jobs}')
 def test_fit_in_sample_residuals_stored_XGBRegressor(n_jobs):
     """
@@ -129,8 +129,8 @@ def test_fit_in_sample_residuals_stored_XGBRegressor(n_jobs):
                            'l2': pd.Series(np.arange(10))})
     
     forecaster = ForecasterAutoregMultiVariate(XGBRegressor(random_state=123),  
-                                               level='l2', lags=3, steps=2)
-    forecaster.fit(series=series, n_jobs=n_jobs)
+                                               level='l2', lags=3, steps=2, n_jobs=n_jobs)
+    forecaster.fit(series=series)
     expected = {1: np.array([-7.98702240e-05, -7.86781311e-05, -9.74178314e-04, 
                               1.15251541e-03, -1.14297867e-03,  1.12581253e-03]),
                 2: np.array([-0.00083256,  0.00097084, -0.00123358,  
@@ -143,7 +143,7 @@ def test_fit_in_sample_residuals_stored_XGBRegressor(n_jobs):
     assert all(all(np.isclose(results[k], expected[k])) for k in expected.keys())
 
 
-@pytest.mark.parametrize("n_jobs", [1, -1], 
+@pytest.mark.parametrize("n_jobs", [1, -1, 'auto'], 
                          ids=lambda n_jobs: f'n_jobs: {n_jobs}')
 def test_fit_same_residuals_when_residuals_greater_than_1000(n_jobs):
     """
@@ -153,14 +153,14 @@ def test_fit_same_residuals_when_residuals_greater_than_1000(n_jobs):
     series = pd.DataFrame({'l1': pd.Series(np.arange(1200)), 
                            'l2': pd.Series(np.arange(1200))})
     
-    forecaster = ForecasterAutoregMultiVariate(LinearRegression(),  
-                                               level='l1', lags=3, steps=2)
-    forecaster.fit(series=series, n_jobs=n_jobs)
+    forecaster = ForecasterAutoregMultiVariate(LinearRegression(), level='l1', 
+                                               lags=3, steps=2, n_jobs=n_jobs)
+    forecaster.fit(series=series)
     results_1 = forecaster.in_sample_residuals
 
     forecaster = ForecasterAutoregMultiVariate(LinearRegression(),  
                                                level='l1', lags=3, steps=2)
-    forecaster.fit(series=series, n_jobs=n_jobs)
+    forecaster.fit(series=series)
     results_2 = forecaster.in_sample_residuals
 
     assert isinstance(results_1, dict)
@@ -173,7 +173,7 @@ def test_fit_same_residuals_when_residuals_greater_than_1000(n_jobs):
     assert all(all(results_1[k] == results_2[k]) for k in results_2.keys())
 
 
-@pytest.mark.parametrize("n_jobs", [1, -1], 
+@pytest.mark.parametrize("n_jobs", [1, -1, 'auto'], 
                          ids=lambda n_jobs: f'n_jobs: {n_jobs}')
 def test_fit_in_sample_residuals_not_stored(n_jobs):
     """
@@ -183,9 +183,9 @@ def test_fit_in_sample_residuals_not_stored(n_jobs):
     series = pd.DataFrame({'l1': pd.Series(np.arange(5)), 
                            'l2': pd.Series(np.arange(5))})
 
-    forecaster = ForecasterAutoregMultiVariate(LinearRegression(),  
-                                               level='l1', lags=3, steps=2)
-    forecaster.fit(series=series, store_in_sample_residuals=False, n_jobs=n_jobs)
+    forecaster = ForecasterAutoregMultiVariate(LinearRegression(), level='l1', 
+                                               lags=3, steps=2, n_jobs=n_jobs)
+    forecaster.fit(series=series, store_in_sample_residuals=False)
     expected = {1: None, 2: None}
     results = forecaster.in_sample_residuals